Fix last author always being registered as unknown
The author_list aggregation will format a list of authors once it
detects that the `id` of the author changes compared to the `id` of the
item. For the last author this never triggers, resulting in the author
not getting registered into the `authors` dict.

The end result is the paper with the most recently added author being
counted as `Unknown`.

This patch changes the aggregation to use a `defaultdict` instead, which
simplifies the code a bit but more importantly removes the bug.